Fix interpolation
ST's example code uses the LUT as is, so it automatically loops back to the
correct point, but we use the LUT and interpolate into another table.  We
therefore need the end of our LUT to also include the start. Without this,
there was a vertical step at the end of each waveform.

(I think there's still a problem here with it now being a single point step
sideways, instead of vertically, but it's significantly better than before)